pyproject.toml
[tool.poetry]
name = "etl-currency-exchange-rates"
version = "0.1.0"
description = ""
authors = ["Renata Hlavova <hlavova.renata@gmail.com>"]


[tool.poetry.group.dev.dependencies]
black = ">=22.10.0"
pylint = ">=2.15.5"

poethepoet = ">=0.16.4"
isort = ">=5.10.1"

[tool.black]
line-length = 120

[tool.poe.tasks]
isort = "isort etl_currency_exchange_rates/ main.py"
isort-check = "isort --check etl_currency_exchange_rates/ main.py"
black = "black etl_currency_exchange_rates/ main.py"
black-check = "black --check etl_currency_exchange_rates/ main.py"

pylint = "pylint etl_currency_exchange_rates/ main.py"

format = ["isort", "black"]
lint = ["isort-check", "black-check", "pylint"]

run = "python -um main"

[tool.poetry.dependencies]
python = "^3.9,<=3.11"
requests = "^2.31.0"
backoff = "^2.2.1"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
